# docs/beacon_sso_sdk_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Beacon SSO SDK 聚合文档，用于 SDK 内，内置 handler 方法展示。",
        "title": "Beacon SSO SDK 聚合文档",
        "contact": {},
        "version": "v1.0.0"
    },
    "paths": {
...
        "/user/by-id": {
            "get": {
                "description": "根据用户ID获取指定用户的详细信息，包括基础信息、联系方式、验证状态和角色列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[用户] 根据ID获取用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID（雪花ID格式）",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pb.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权或令牌失效",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/userinfo": {
            "get": {
                "description": "通过访问令牌获取当前登录用户的详细信息，包括基础信息、联系方式、验证状态和角色列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "[玩家] 用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pb.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权或令牌失效",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "pb.Role": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "角色代码（如 SUPER_ADMIN、ADMIN、MERCHANT）",
                    "type": "string"
                },
                "description": {
                    "description": "角色描述",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                }
            }
        },
        "pb.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像 URL",
                    "type": "string"
                },
                "birthday": {
                    "description": "生日（ISO 8601 日期格式）",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱地址",
                    "type": "string"
                },
                "gender": {
                    "description": "性别（0 未知 / 1 男 / 2 女）",
                    "type": "integer"
                },
                "id": {
                    "description": "用户 ID（雪花 ID 格式）",
                    "type": "string"
                },
                "is_email_verified": {
                    "description": "邮箱是否已验证",
                    "type": "boolean"
                },
                "is_phone_verified": {
                    "description": "手机是否已验证",
                    "type": "boolean"
                },
                "last_login_at": {
                    "description": "最后登录时间（ISO 8601 格式）",
                    "type": "string"
                },
                "last_login_ip": {
                    "description": "最后登录 IP",
                    "type": "string"
                },
                "need_reset_password": {
                    "description": "是否需要强制重置密码",
                    "type": "boolean"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号码",
                    "type": "string"
                },
                "roles": {
                    "description": "用户角色列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Role"
                    }
                },
                "status": {
                    "description": "状态（0 禁用 / 1 启用）",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "xBase.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "string"
                },
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "overhead": {
                    "type": "integer"
                }
            }
        },
...
    }
}